pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "moodlectl"
version = "0.1.0"
description = "Moodle LMS automation and AI tool for university instructors"
requires-python = ">=3.12"

dependencies = [
    "requests>=2.33",
    "beautifulsoup4>=4.14",
    "typer>=0.15",
    "rich>=13",
    "python-dotenv>=1.0",
    "anthropic>=0.40",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-mock>=3.14",
    "responses>=0.25",
]
export = [
    "openpyxl>=3.1",
]

[project.scripts]
moodlectl = "moodlectl.cli.main:app"

[tool.ruff]
target-version = "py312"
line-length = 100
